refactor(terminal): remove automatic performance mode
Remove automatic performance mode triggering based on terminal count. Performance mode is now manual-only toggle in Settings.

- Delete useTerminalPerformance hook and auto-enable logic
- Simplify performanceModeStore to single boolean state
- Centralize data-performance-mode attribute management in AppLayout
- Remove auto-enable UI (threshold input, terminal count display)
- Update Settings UI copy to describe manual safe mode
- Remove performanceModeAutoEnableThreshold from IPC types
- Delegate DOM attribute handling from stateHydration to AppLayout

Performance mode remains available as manual toggle for low-end hardware or user preference. With grid cap (16), docking, WebGL budgeting, and per-type scrollback policies, automatic triggering provides minimal value.
